move vpromms row properties to own file
